requirements.txt
Flask
smtplib
email
PyPDF2
reportlab
requests
python-dotenv
